PERA-2609 Implement swap no account state
