+++ app/src/main/kotlin/com/algorand/android/ui/swap/topfive/viewmodel/DefaultTopSwapPairsViewModel.kt
@@ -31,6 +31,7 @@ import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.combine
 import kotlinx.coroutines.flow.launchIn
+import kotlinx.coroutines.flow.onEach
 import kotlinx.coroutines.launch
 
 @HiltViewModel
@@ -60,6 +61,15 @@ class DefaultTopSwapPairsViewModel @Inject constructor(
                 } else {
                     stateDelegate.updateState { ViewState.Idle }
                 }
+            }.launchIn(viewModelScope)
+        }
+    }
+
+    override fun init() {
+        stateDelegate.onState<ViewState.Idle> {
+            initTopFiveSwaps()
+            viewStateFlow.onEach { viewState ->
+                stateDelegate.updateState { viewState }
             }.launchIn(viewModelScope)
         }
     }
