+++ app/src/main/kotlin/com/algorand/android/ui/swap/topfive/viewmodel/TopSwapPairsViewModel.kt
@@ -22,6 +22,8 @@ interface TopSwapPairsViewModel : StateViewModel<ViewState> {
 
     fun init(swapWidgetViewState: Flow<SwapWidgetViewModel.ViewState>)
 
+    fun init()
+
     sealed interface ViewState {
         data object Idle : ViewState
         data object Loading : ViewState
