+++ app/src/main/res/values/strings.xml
@@ -1557,4 +1557,6 @@
     <string name="asset_to_asset_formatted">%1$s to %2$s</string>
     <string name="unable_to_load_top_five">Unable to load Top 5 Swaps</string>
     <string name="swap_successfully_completed_formatted">%1$s to %2$s swap successfully completed</string>
+    <string name="create_an_account_to_get">Create an account to get swapping</string>
+    <string name="create_account">Create Account</string>
 </resources>
